requirements.txt
Django==1.7.4
Markdown==2.5.2
MySQL-python==1.2.5
Pillow==2.7.0
Pygments==2.0.2
South==1.0.2
argparse==1.2.1
django-filter==0.9.2
djangorestframework==3.0.4
wsgiref==0.1.2
